Universe correctly calcs coordinates

+++ src/Region.java
@@ -12,8 +12,8 @@ public class Region {
         this.name = name;
     }
 
-    public double findDistance(Player player) {
-        return Math.sqrt(Math.pow(this.x - player.getRegion().getX(), 2) +
+    public int findDistance(Player player) {
+        return (int) Math.sqrt(Math.pow(this.x - player.getRegion().getX(), 2) +
                 Math.pow(this.y - player.getRegion().getY(), 2));
     }
